backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
aiofiles>=23.2.1
httpx>=0.26.0
